Use database to store lunch information
A MongoDB Atlas now stores the information regarding the lunch
start date and end date, and the resume work reminders.
One of the reasons was to store the lunch information somewhere,
but mainly to save the reminders so that they can be later marked
as deleted on Slack.

There is some difficulty on how to deleted these reminders. There
is a timeout to delete these reminders. If the lunch exceeds 30
minutes, the Heroku turns off the dyno running the application,
thus not being able to execute the timeout.
The resorted solution was to store the reminders in the database,
to later execute a scheduled routine to delete them.